add service controller and service

+++ backend/Server/Models/DatabaseObjects/Service.cs
@@ -2,7 +2,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace backend.Server.Models.DatabaseObjects;
 public class Service
-{
+    {
     [Key]
     public long Nid { get; set; }
     public string Name { get; set; }
@@ -11,5 +11,5 @@ public class Service
     public long VatId { get; set; }
     public DateTime? DiscountTime { get; set; }
     public int TimeMin { get; set; }
-    public int BusinessId { get; set; }
-}
+    public long BusinessId { get; set; }
+    }
